+++ integ/components/deadline/deadline_01_repository/bin/deadline_01_repository.ts
@@ -4,6 +4,11 @@
  */
 
 import { App, Stack } from '@aws-cdk/core';
+import {
+  Stage,
+  ThinkboxDockerRecipes,
+} from 'aws-rfdk/deadline';
+
 import { DatabaseType, StorageStruct } from '../../../../lib/storage-struct';
 import { RepositoryTestingTier } from '../lib/repository-testing-tier';
 
@@ -17,17 +22,26 @@ const integStackTag = process.env.INTEG_STACK_TAG!.toString();
 
 const componentTier = new Stack(app, 'RFDKInteg-DL-ComponentTier' + integStackTag, {env});
 
+const stagePath = process.env.DEADLINE_STAGING_PATH!.toString();
+// Stage docker recipes, which include the repo installer in (`recipes.version`)
+const recipes = new ThinkboxDockerRecipes(componentTier, 'DockerRecipes', {
+  stage: Stage.fromDirectory(stagePath),
+});
+
 const structs: Array<StorageStruct> = [
   new StorageStruct(componentTier, 'StorageStruct1', {
     integStackTag,
+    version: recipes.version,
   }),
   new StorageStruct(componentTier, 'StorageStruct2', {
     integStackTag,
     databaseType: DatabaseType.DocDB,
+    version: recipes.version,
   }),
   new StorageStruct(componentTier, 'StorageStruct3', {
     integStackTag,
     databaseType: DatabaseType.MongoDB,
+    version: recipes.version,
   }),
 ];