+++ integ/scripts/bash/rfdk-integ-e2e.sh
@@ -71,13 +71,14 @@ export INFRASTRUCTURE_DEPLOY_FINISH_TIME=$SECONDS
 
 # Pull the top level directory for each cdk app in the components directory
 for COMPONENT in **/cdk.json; do
+    # In case the yarn install was done inside this integ package, there are some example cdk.json files in the aws-cdk
+    # package we want to avoid.
+    if [[ $COMPONENT == *"node_modules"* ]]; then
+        continue
+    fi
+
     COMPONENT_ROOT="$(dirname "$COMPONENT")"
     COMPONENT_NAME=$(basename "$COMPONENT_ROOT")
-    # Invoke hook function if it is exported and name is defined in PRE_COMPONENT_HOOK variable
-    if [ -n "$PRE_COMPONENT_HOOK" ]  && [ "$(type -t $PRE_COMPONENT_HOOK)" == "function" ]
-    then
-      $PRE_COMPONENT_HOOK $COMPONENT_NAME
-    fi
 
     # Use a pattern match to exclude the infrastructure app from the results
     export ${COMPONENT_NAME}_START_TIME=$SECONDS