+++ integ/scripts/bash/stage-deadline.sh
@@ -8,7 +8,7 @@ set -euo pipefail
 # Check if Deadline assets are staged at configured path
 if [ ! $(ls "$DEADLINE_STAGING_PATH/manifest.json" 2> /dev/null) ]; then
     # Stage Deadline assets
-    npx ../packages/aws-rfdk/bin/stage-deadline --output "$DEADLINE_STAGING_PATH" $DEADLINE_VERSION
+    npx stage-deadline --output "$DEADLINE_STAGING_PATH" $DEADLINE_VERSION
 fi
 
 exit 0