+++ packages/aws-rfdk/lib/deadline/lib/index.ts
@@ -13,4 +13,6 @@ export * from './render-queue-ref';
 export * from './stage';
 export * from './thinkbox-docker-recipes';
 export * from './version';
+export * from './version-query';
 export * from './version-ref';
+export * from './worker-configuration';